RDKTV-14897: Fixed race condition in deleting route available flag.
Reason for change: Updated condition to check for all the gateways
                   to be empty before deleting  route available flag.

Test Procedure: 1. Connect device in dual stack mode(both ipv6 and ipv4).
                2. Disable ipv6, amd make sure the flag is still present.

Risks: Low.

(cherry picked from commit c6a1b7320dd08221af81200ae721af1119adf099)

+++ checkDefaultRoute.sh
@@ -63,7 +63,7 @@ elif [ "$opern" = "delete" ]; then
     #Remove flag and IP for delete operation
     echo "Deleting Route Flag"
     sed -i "/$gtwip/d" $FILE
-    rm -rf /tmp/route_available
+    [ -s $FILE ] || rm -rf /tmp/route_available
 else
     echo "Received operation:$opern is Invalid..!!"
 fi
